feat: Add MIT License
This commit introduces the MIT License to the project. The license
grants permission to any person obtaining a copy of this software and
associated documentation files to deal in the Software without
restriction. This includes rights to use, copy, modify, merge, publish,
distribute, sublicense, and/or sell copies of the Software.